.cosmetics
git-svn-id: svn://ultimatepp.org/upp/trunk@6828 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Stream.cpp
@@ -1534,7 +1534,7 @@ String LoadStream(Stream& in) {
 		int64 size = in.GetLeft();
 		if(size >= 0 && size < INT_MAX / 2) {
 			StringBuffer s((int)size);
-			in.Get(s, size);
+			in.Get(s, (int)size);
 			if(!in.IsError())
 				return s;
 		}
